requirements.txt
aioredis==1.3.1
ansicon==1.89.0
appdirs==1.4.4
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
blessed==1.19.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-access-policy==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jinxed==1.1.0
msgpack==1.0.3
mysqlclient==2.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
wcwidth==0.2.5
zope.interface==5.4.0

requests~=2.27.1
environ~=1.0